Refactor(exporter): decouple TraceCommittee from HTTP framework
